requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
moviepy==1.0.3
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
decorator>=4.0.2,<5.0
proglog==0.1.10
tqdm==4.66.1
requests==2.31.0
openai==1.35.7
# New dependencies for auth and db
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
rich
